+++ backend/src/auth/auth.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Get, Post, Headers, UseGuards, Param, Patch } from "@nestjs/common";
+import { Body, Controller, Get, Post, Headers, UseGuards, Patch } from "@nestjs/common";
 import { CreateUserDto } from "../user/dto/create-user.dto";
 import { AuthService } from "./auth.service";
 import { LoginUserDto } from "../user/dto/login-user.dto";
@@ -6,8 +6,9 @@ import { ApiOperation, ApiTags, ApiOkResponse } from "@nestjs/swagger";
 import { UserModel } from "../user/models/user.model";
 import { RolesGuard } from "./guards/roles.guard";
 import { Roles } from "./decorators/role-auth.decorator";
-import { ForgotPasswordDto } from "../user/dto/forgot-password.dto";
-import { ChangePasswordDto } from "../user/dto/change-password.dto";
+import { SendVerificationEmailDto } from "../user/dto/send-verification-email.dto";
+import { SetPasswordDto } from "../user/dto/set-password.dto";
+import { VerifyCodeDto } from "../user/dto/verify-code.dto";
 
 @ApiTags("Auth functional")
 @Controller("auth")
@@ -43,18 +44,24 @@ export class AuthController {
     return this.authService.me(headers.authorization);
   }
 
-  @ApiOperation({ summary: "Forgot password request" })
+  @ApiOperation({ summary: "Send verification email" })
   @ApiOkResponse({ status: 200, type: UserModel })
-  @Post("/forgot_password")
-  forgotPassword(@Body() dto: ForgotPasswordDto) {
-    console.log(dto);
-    return this.authService.forgotPassword(dto.email);
+  @Post("/send_verification_email")
+  sendVerificationEmail(@Body() dto: SendVerificationEmailDto): Promise<void> {
+    return this.authService.sendVerificationEmail(dto.email);
+  }
+
+  @ApiOperation({ summary: "Verify code" })
+  @ApiOkResponse({ status: 200, type: UserModel })
+  @Patch("/verify_code")
+  verifyCode(@Body() dto: VerifyCodeDto): Promise<boolean> {
+    return this.authService.verifyCode(dto.email, dto.code);
   }
 
   @ApiOperation({ summary: "Change password" })
   @ApiOkResponse({ status: 200, type: UserModel })
-  @Patch("/change_password/:token")
-  changePassword(@Body() dto: ChangePasswordDto, @Param("token") token: string) {
-    return this.authService.changePassword(token, dto);
+  @Patch("/set_password")
+  changePassword(@Body() dto: SetPasswordDto): Promise<boolean> {
+    return this.authService.changePassword(dto.email, dto.code, dto.password);
   }
 }