+++ frontend/src/types/auth/set-password.dto.ts
@@ -1,0 +1,5 @@
+import { VerifyCodeDto } from './verify-code.dto';
+
+export interface SetPasswordDto extends VerifyCodeDto{
+  password: string;
+}